requirements.txt
Django==4.0.6
djangorestframework
djangorestframework-simplejwt
django-environ
django-storages
django-cors-headers
django-braces
django-extensions
psycopg2-binary
arrow
gunicorn
pytz